Fix web3 ethAdapter tests
